:recycle: Move should_print_* verbosity gates into command::version
