+++ src/main.rs
@@ -1,8 +1,7 @@
 extern crate cargo_semantic_release;
-use cargo_semantic_release::{RepositoryExtension, SemanticReleaseConfig, SemanticVersionAction};
+use cargo_semantic_release::SemanticVersionAction;
 use clap::Parser;
 use clap_cargo::style;
-use semver::Version;
 
 mod command;
 mod undo_state;
@@ -115,145 +114,5 @@ fn main() {
             command::version::run_version_command(version_args, verbosity, noop)
         }
         SemanticReleaseCommand::Undo(undo_args) => command::undo::run_undo_command(undo_args, noop),
-    }
-}
-
-/// Whether the commits since the last version tag should be printed, given `verbosity`.
-///
-/// Requires `-vv` (or higher); `-v` alone does not show the commit log.
-fn should_print_commit_log(verbosity: u8) -> bool {
-    verbosity >= 2
-}
-
-/// Whether the `Cargo.toml` version should be printed, given `verbosity`.
-///
-/// Requires `-v` (or higher).
-fn should_print_cargo_toml_version(verbosity: u8) -> bool {
-    verbosity >= 1
-}
-
-/// Whether the found tags should be printed, given `verbosity`.
-///
-/// Requires `-v` (or higher).
-fn should_print_found_tags(verbosity: u8) -> bool {
-    verbosity >= 1
-}
-
-/// Whether the latest tag version should be printed, given `verbosity`.
-///
-/// Requires `-v` (or higher).
-fn should_print_latest_tags_version(verbosity: u8) -> bool {
-    verbosity >= 1
-}
-
-/// The repository's current version: the latest version tag's version, or `0.0.0` if there is
-/// none yet.
-fn current_version(
-    repository: &impl RepositoryExtension,
-    config: &SemanticReleaseConfig,
-) -> Result<Version, Box<dyn std::error::Error>> {
-    Ok(repository
-        .get_latest_version_tag(&config.tag_format)?
-        .map(|tag| tag.version)
-        .unwrap_or_else(|| Version::new(0, 0, 0)))
-}
-
-#[cfg(test)]
-mod should_print_commit_log_tests {
-    use crate::should_print_commit_log;
-
-    #[test]
-    fn is_false_below_double_verbose() {
-        assert!(!should_print_commit_log(0));
-        assert!(!should_print_commit_log(1));
-    }
-
-    #[test]
-    fn is_true_at_double_verbose_or_more() {
-        assert!(should_print_commit_log(2));
-        assert!(should_print_commit_log(3));
-    }
-}
-
-#[cfg(test)]
-mod should_print_cargo_toml_version_tests {
-    use crate::should_print_cargo_toml_version;
-
-    #[test]
-    fn is_false_without_verbose() {
-        assert!(!should_print_cargo_toml_version(0));
-    }
-
-    #[test]
-    fn is_true_at_verbose_or_more() {
-        assert!(should_print_cargo_toml_version(1));
-        assert!(should_print_cargo_toml_version(2));
-    }
-}
-
-#[cfg(test)]
-mod should_print_found_tags_tests {
-    use crate::should_print_found_tags;
-
-    #[test]
-    fn is_false_without_verbose() {
-        assert!(!should_print_found_tags(0));
-    }
-
-    #[test]
-    fn is_true_at_verbose_or_more() {
-        assert!(should_print_found_tags(1));
-        assert!(should_print_found_tags(2));
-    }
-}
-
-#[cfg(test)]
-mod should_print_latest_tags_version_tests {
-    use crate::should_print_latest_tags_version;
-
-    #[test]
-    fn is_false_without_verbose() {
-        assert!(!should_print_latest_tags_version(0));
-    }
-
-    #[test]
-    fn is_true_at_verbose_or_more() {
-        assert!(should_print_latest_tags_version(1));
-        assert!(should_print_latest_tags_version(2));
-    }
-}
-
-#[cfg(test)]
-mod current_version_tests {
-    use crate::current_version;
-    use cargo_semantic_release::test_util::{repo_init, RepositoryTestExtensions};
-    use cargo_semantic_release::SemanticReleaseConfig;
-    use semver::Version;
-
-    #[test]
-    fn defaults_to_0_0_0_without_a_version_tag() {
-        // Given
-        let (_temp_dir, repository) = repo_init(None);
-
-        // When
-        let result = current_version(&repository, &SemanticReleaseConfig::default());
-
-        // Then
-        assert_eq!(result.unwrap(), Version::new(0, 0, 0));
-    }
-
-    #[test]
-    fn reads_the_version_from_the_latest_version_tag() {
-        // Given
-        let commit_message = ":tada: initial release";
-        let (_temp_dir, repository) = repo_init(Some(vec![commit_message]));
-        let tagged_commit = repository.find_commit_by_message(commit_message).unwrap();
-        repository.add_tag(tagged_commit, "v1.2.3");
-
-        // When
-        let result = current_version(&repository, &SemanticReleaseConfig::default());
-
-        // Then
-        assert_eq!(result.unwrap(), Version::new(1, 2, 3));
     }
 }
